+++ src/messages/layouts/MessageLayout.cpp
@@ -328,6 +328,11 @@ void MessageLayout::updateBuffer(QPixmap *buffer, int /*messageIndex*/,
     {
         backgroundColor = QColor("#4A273D");
     }
+    else if (this->message_->flags.has(MessageFlag::WebchatDetected) &&
+             getSettings()->normalNonceDetection)
+    {
+        backgroundColor = blendColors(backgroundColor, QColor("#3FFFA30B"));
+    }
 
     painter.fillRect(buffer->rect(), backgroundColor);